iojs-*: cp -a into /build/iojs to avoid permission issues
If the build is executed directly in the mounted host-volume then it
will fail if the user on the host machine is not the same than the user
in the container (which is most likely true all the time).

Plus, this way we avoid to leave executables on the host-machine. They
are built on the container and removed with the container when the
container ends building it.